Set version to 2.1.3, removing beta. Add arm64 slice to simulator fat binary.
