+++ README.md
@@ -1,5 +1,5 @@
 # TapResearch-iOS-SDK
-TapResearch iOS SDK v2.1.3-beta2
+TapResearch iOS SDK v2.1.3
 
 
 For additional information, please see the [TapResearch iOS SDK integration guide](https://www.tapresearch.com/docs/ios_integration_guide).
@@ -13,7 +13,7 @@ Create an [app](/supplier_dashboard/dashboard/apps/new) and grab your API Token.
 
 Add the pod info to the app's Podfile
  ```ruby
- pod 'TapResearch', '2.1.3-beta2'
+ pod 'TapResearch', '2.1.3'
  ```
 
  Then run the following command:
