+++ TapResearchSDK.xcframework/ios-arm64_i386_x86_64-simulator/TapResearchSDK.framework/Headers/TRPlacementCustomParameter.h
@@ -1,0 +1,20 @@
+//
+//  TRPlacementParameter.h
+//  TapResearchSDK
+//
+//  Created by Ilan Caspi on 1/16/20.
+//  Copyright © 2020 TapResearch. All rights reserved.
+//
+
+#import <Foundation/Foundation.h>
+
+NS_ASSUME_NONNULL_BEGIN
+
+@interface TRPlacementCustomParameter : NSObject
+
+@property (nonatomic, strong) NSString *key;
+@property (nonatomic, strong) NSString *value;
+
+@end
+
+NS_ASSUME_NONNULL_END
