+++ TapResearchSDK.xcframework/ios-arm64_i386_x86_64-simulator/TapResearchSDK.framework/Modules/module.modulemap
@@ -1,0 +1,6 @@
+framework module TapResearchSDK {
+  umbrella header "TapResearchSDK.h"
+
+  export *
+  module * { export * }
+}
