--- TapResearchSDK.xcframework/ios-i386_x86_64-simulator/TapResearchSDK.framework/_CodeSignature/CodeResources
@@ -1,252 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<!DOCTYPE plist PUBLIC "-//Apple//DTD PLIST 1.0//EN" "http://www.apple.com/DTDs/PropertyList-1.0.dtd">
-<plist version="1.0">
-<dict>
-	<key>files</key>
-	<dict>
-		<key>Headers/TRPlacement.h</key>
-		<data>
-		j63gV+2UUu5TCQ11kmuxCLI1LCY=
-		</data>
-		<key>Headers/TRPlacementCustomParameter+Builder.h</key>
-		<data>
-		KXh/ELCZwDxV1fwKz9asZSMFODg=
-		</data>
-		<key>Headers/TRPlacementCustomParameter.h</key>
-		<data>
-		5POiPUmUsk0TSBWzsnAol5LgpJk=
-		</data>
-		<key>Headers/TRPlacementCustomParameterBuilder.h</key>
-		<data>
-		gVZZt6d9/4wkrn/HyJ354LfXwMQ=
-		</data>
-		<key>Headers/TRPlacementCustomParameterList.h</key>
-		<data>
-		kC/1/n+STeh1Ge84ZDI18bbE93Y=
-		</data>
-		<key>Headers/TRReward.h</key>
-		<data>
-		Q/mt+0sz7OTZRwsHsn2ZK5xbXxU=
-		</data>
-		<key>Headers/TRSerializationHelper.h</key>
-		<data>
-		eoxCllwC+fLLH2F4zByBXg/QAkw=
-		</data>
-		<key>Headers/TapResearch.h</key>
-		<data>
-		vQw0Z2Y0iNWPISz+B7JrZrcbnqM=
-		</data>
-		<key>Headers/TapResearchSDK.h</key>
-		<data>
-		rB5EkYbDKnW8KJtKUZgsfCycPIk=
-		</data>
-		<key>Info.plist</key>
-		<data>
-		cG5hc4oaDW1m3psGg+PGgqCvDYg=
-		</data>
-		<key>Modules/module.modulemap</key>
-		<data>
-		t2mK6e1jqz4PMwXFB4jsTlu8hno=
-		</data>
-	</dict>
-	<key>files2</key>
-	<dict>
-		<key>Headers/TRPlacement.h</key>
-		<dict>
-			<key>hash</key>
-			<data>
-			j63gV+2UUu5TCQ11kmuxCLI1LCY=
-			</data>
-			<key>hash2</key>
-			<data>
-			h8hAAENQPFy8TF0XX+eCVKKWFisOaBs2kJcfZLOztHo=
-			</data>
-		</dict>
-		<key>Headers/TRPlacementCustomParameter+Builder.h</key>
-		<dict>
-			<key>hash</key>
-			<data>
-			KXh/ELCZwDxV1fwKz9asZSMFODg=
-			</data>
-			<key>hash2</key>
-			<data>
-			eQj0VHcQj31pyjnHj8YB010AUzRQzBkaN6Mn5y2UmNA=
-			</data>
-		</dict>
-		<key>Headers/TRPlacementCustomParameter.h</key>
-		<dict>
-			<key>hash</key>
-			<data>
-			5POiPUmUsk0TSBWzsnAol5LgpJk=
-			</data>
-			<key>hash2</key>
-			<data>
-			6rM5Q02P2rkxQbIA+8FbtKZlcZS7TgBRAfSJyoMsxt4=
-			</data>
-		</dict>
-		<key>Headers/TRPlacementCustomParameterBuilder.h</key>
-		<dict>
-			<key>hash</key>
-			<data>
-			gVZZt6d9/4wkrn/HyJ354LfXwMQ=
-			</data>
-			<key>hash2</key>
-			<data>
-			ijH9feR8/TT4BaKSQ1SOkgMbAjulBEd3GPMtl+W3HrY=
-			</data>
-		</dict>
-		<key>Headers/TRPlacementCustomParameterList.h</key>
-		<dict>
-			<key>hash</key>
-			<data>
-			kC/1/n+STeh1Ge84ZDI18bbE93Y=
-			</data>
-			<key>hash2</key>
-			<data>
-			kHB+FFoKwOmea6r6QGgNed1JgKzuIMLtOVK3tZvJeQE=
-			</data>
-		</dict>
-		<key>Headers/TRReward.h</key>
-		<dict>
-			<key>hash</key>
-			<data>
-			Q/mt+0sz7OTZRwsHsn2ZK5xbXxU=
-			</data>
-			<key>hash2</key>
-			<data>
-			Ht5HCgEiIQFYYAqpBM+w6nW2+tTX5/TbphVjp1cleyA=
-			</data>
-		</dict>
-		<key>Headers/TRSerializationHelper.h</key>
-		<dict>
-			<key>hash</key>
-			<data>
-			eoxCllwC+fLLH2F4zByBXg/QAkw=
-			</data>
-			<key>hash2</key>
-			<data>
-			h7DvUmVisKssm3gKx9HFfuUMkbwC/+lnf0sTQG6p9mY=
-			</data>
-		</dict>
-		<key>Headers/TapResearch.h</key>
-		<dict>
-			<key>hash</key>
-			<data>
-			vQw0Z2Y0iNWPISz+B7JrZrcbnqM=
-			</data>
-			<key>hash2</key>
-			<data>
-			JiCHlkCV4VSsGOSXjfR2Wh7LTSQV3RSufKb4aoc/BdM=
-			</data>
-		</dict>
-		<key>Headers/TapResearchSDK.h</key>
-		<dict>
-			<key>hash</key>
-			<data>
-			rB5EkYbDKnW8KJtKUZgsfCycPIk=
-			</data>
-			<key>hash2</key>
-			<data>
-			90lmrgaQESctQag9B7Iop/WJLHW9dTrB3IJ1Xmwl0Xk=
-			</data>
-		</dict>
-		<key>Modules/module.modulemap</key>
-		<dict>
-			<key>hash</key>
-			<data>
-			t2mK6e1jqz4PMwXFB4jsTlu8hno=
-			</data>
-			<key>hash2</key>
-			<data>
-			/ofKK6zgGzETCg0ME5RFie2rVeRv1XLd8oI7FJDsKrk=
-			</data>
-		</dict>
-	</dict>
-	<key>rules</key>
-	<dict>
-		<key>^.*</key>
-		<true/>
-		<key>^.*\.lproj/</key>
-		<dict>
-			<key>optional</key>
-			<true/>
-			<key>weight</key>
-			<real>1000</real>
-		</dict>
-		<key>^.*\.lproj/locversion.plist$</key>
-		<dict>
-			<key>omit</key>
-			<true/>
-			<key>weight</key>
-			<real>1100</real>
-		</dict>
-		<key>^Base\.lproj/</key>
-		<dict>
-			<key>weight</key>
-			<real>1010</real>
-		</dict>
-		<key>^version.plist$</key>
-		<true/>
-	</dict>
-	<key>rules2</key>
-	<dict>
-		<key>.*\.dSYM($|/)</key>
-		<dict>
-			<key>weight</key>
-			<real>11</real>
-		</dict>
-		<key>^(.*/)?\.DS_Store$</key>
-		<dict>
-			<key>omit</key>
-			<true/>
-			<key>weight</key>
-			<real>2000</real>
-		</dict>
-		<key>^.*</key>
-		<true/>
-		<key>^.*\.lproj/</key>
-		<dict>
-			<key>optional</key>
-			<true/>
-			<key>weight</key>
-			<real>1000</real>
-		</dict>
-		<key>^.*\.lproj/locversion.plist$</key>
-		<dict>
-			<key>omit</key>
-			<true/>
-			<key>weight</key>
-			<real>1100</real>
-		</dict>
-		<key>^Base\.lproj/</key>
-		<dict>
-			<key>weight</key>
-			<real>1010</real>
-		</dict>
-		<key>^Info\.plist$</key>
-		<dict>
-			<key>omit</key>
-			<true/>
-			<key>weight</key>
-			<real>20</real>
-		</dict>
-		<key>^PkgInfo$</key>
-		<dict>
-			<key>omit</key>
-			<true/>
-			<key>weight</key>
-			<real>20</real>
-		</dict>
-		<key>^embedded\.provisionprofile$</key>
-		<dict>
-			<key>weight</key>
-			<real>20</real>
-		</dict>
-		<key>^version\.plist$</key>
-		<dict>
-			<key>weight</key>
-			<real>20</real>
-		</dict>
-	</dict>
-</dict>
-</plist>
